+++ packages/backend/src/model/denormalize/index.js
@@ -1,0 +1,3 @@
+const denormalizeScene = require("./denormalizeScene");
+
+module.exports = { denormalizeScene };
