+++ packages/backend/src/model/denormalize/utils/index.js
@@ -1,0 +1,4 @@
+const denormalizeNumber = require("./denormalizeNumber");
+const denormalizeBonanza = require("./denormalizeBonanza");
+
+module.exports = { denormalizeNumber, denormalizeBonanza };
